requirements.txt
robotframework-seleniumlibrary==4.1.0
robotframework==3.2.1
robotframework-faker==4.3.0
robotframework-pabot==1.2.1
robotframework-datadriver==0.3.6
robotframework-requests= 0.7.0
selenium==3.141.0
pydevd==1.9.0
psutil==5.7.0
pywin32==227